backend-tally-dashboard/requirements.txt
# Django Core
Django==5.2
psycopg2-binary==2.9.11
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
gevent==24.11.1

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==24.2

# Configuration Management
python-decouple==3.8

# Data Processing
pandas==2.3.3
numpy==2.3.4
openpyxl==3.1.5
et-xmlfile==2.0.0

# Date/Time Utilities
pytz==2023.3
python-dateutil==2.9.0.post0
tzdata==2025.2

# Email Service (ZeptoMail API)
requests==2.32.3

# Cloud Storage (Cloudflare R2 / S3)
django-storages==1.14.2
boto3==1.35.0

# Core Dependencies
asgiref==3.10.0
sqlparse==0.5.3
packaging==25.0
six==1.17.0
wcwidth==0.2.14
